+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add new controller methods `autoscale()`, `descale()`, and `started()` to facilitate different kinds of Shiny apps.
 * Deprecate the `scale` and `throttle` methods of `controller$promise()`. `promise()` now always calls `autoscale()` to make sure one and only one auto-scaling loop is running asynchronously. Auto-scaling thus continues even after the promise resolves.
 * Add a second example vignette that simulates coin flips.
+* Add a new `error` argument to `collect()` (#166).
 
 # crew 0.9.1